+++ core/scheduler/app_registry.go
@@ -1,4 +1,4 @@
-package routehttp
+package scheduler
 
 import (
 	"encoding/json"
