+++ core/scheduler/http_test.go
@@ -1,4 +1,4 @@
-package routehttp
+package scheduler
 
 import (
 	"bytes"
